+++ examples/gmd_timings/validation.py
@@ -1,0 +1,9 @@
+import fv3gfs.wrapper
+
+if __name__ == "__main__":
+    fv3gfs.wrapper.initialize()
+    for i in range(fv3gfs.wrapper.get_step_count()):
+        fv3gfs.wrapper.step_dynamics()
+        fv3gfs.wrapper.step_physics()
+        fv3gfs.wrapper.save_intermediate_restart_if_enabled()
+    fv3gfs.wrapper.cleanup()